refactor: move renew policy into subpackage
