+++ policy/renew/get-renew-policy-by-uid-handler.go
@@ -1,4 +1,4 @@
-package policy
+package renew
 
 import (
 	"encoding/json"
